backend/requirements.txt
python-dotenv>=0.19.0
slack_sdk>=3.19.0
